editor/libeditor/tests/test_bug735059.html

<!DOCTYPE html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=735059
-->
<title>Test for Bug 735059</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=735059">Mozilla Bug 735059</a>
<div id="display" contenteditable>foo</div>
<pre id="test">
<script>
/** Test for Bug 735059 **/

// Value defaults to the empty string, which evaluates to true, so this
// disables CSS styling
document.execCommand("usecss");
getSelection().selectAllChildren(document.getElementById("display"));
document.execCommand("bold");
is(document.getElementById("display").innerHTML, "<b>foo</b>",
   "execCommand() needs to work with only one parameter");
</script>
</pre>
